segmented code into multiple files, added program header editing routine. time to try to inject some code
