+++ includes/woody.h
@@ -25,4 +25,13 @@
 
 # include <libft.h>
 
+
+void				print_hdr(void *ptr);
+void				print_shdr(void *ptr);
+void				print_phdr(void *ptr);
+void				print_all(void *ptr);
+
+Elf64_Shdr			*get_section_byindex_64(Elf64_Ehdr *hdr, Elf64_Half index);
+Elf64_Shdr			*get_section_bytype_64(Elf64_Ehdr *hdr, Elf64_Word type);
+
 #endif
