+++ srcs/main.c
@@ -12,97 +12,6 @@
 
 #include <woody.h>
 
-Elf64_Shdr			*get_section_64(Elf64_Ehdr *hdr, Elf64_Half index)
-{
-	Elf64_Shdr		*shdr;
-
-	shdr = (void *)hdr + hdr->e_shoff;
-	for (int i = 0; i < hdr->e_shnum; i++)
-	{
-		if (i == index)
-			return (shdr);
-		shdr = (void *)shdr + sizeof(Elf64_Shdr);
-	}
-	return (NULL);
-}
-
-void				print_all(void *ptr)
-{
-	Elf64_Ehdr		*hdr;
-	Elf64_Shdr		*shdr;
-	Elf64_Shdr		*sh_strtable;
-	char			*strtable;
-	int				i;
-
-	hdr = ptr;
-	printf("Header:\n\te_ident: ");
-	for (i = 0; i < EI_NIDENT; ++i) {
-		printf("%02X ", hdr->e_ident[i]);
-	}
-	printf("\n\te_type: %hu (%hX)", hdr->e_type, hdr->e_type);
-	printf("\n\te_machine: %hu (%hX)", hdr->e_machine, hdr->e_machine);
-	printf("\n\te_version: %u (%X)", hdr->e_version, hdr->e_version);
-	printf("\n\te_entry: %lu (%lX)", hdr->e_entry, hdr->e_entry);
-	printf("\n\te_phoff: %lu (%lX)", hdr->e_phoff, hdr->e_phoff);
-	printf("\n\te_shoff: %lu (%lX)", hdr->e_shoff, hdr->e_shoff);
-	printf("\n\te_flags: %u (%X)", hdr->e_flags, hdr->e_flags);
-	printf("\n\te_ehsize: %hu (%hX)", hdr->e_ehsize, hdr->e_ehsize);
-	printf("\n\te_phentsize: %hu (%hX)", hdr->e_phentsize, hdr->e_phentsize);
-	printf("\n\te_phnum: %hu (%hX)", hdr->e_phnum, hdr->e_phnum);
-	printf("\n\te_shentsize: %hu (%hX)", hdr->e_shentsize, hdr->e_shentsize);
-	printf("\n\te_shnum: %hu (%hX)", hdr->e_shnum, hdr->e_shnum);
-	printf("\n\te_shstrndx: %hu (%hX)\n", hdr->e_shstrndx, hdr->e_shstrndx);
-
-	shdr = (void *)hdr + hdr->e_shoff;
-	// get shstrtable section
-	sh_strtable = get_section_64(hdr, hdr->e_shstrndx);
-	strtable = (void *)hdr + sh_strtable->sh_offset;
-	printf("Sections:\n");
-	for (int i = 0; i < hdr->e_shnum; i++)
-	{
-		printf("\n\n\t%s:\n", strtable + shdr->sh_name);
-
-		printf("sh_name: %u\n", shdr->sh_name);
-		printf("sh_type: %u\n", shdr->sh_type);
-		printf("sh_flags: %lu\n", shdr->sh_flags);
-		printf("sh_addr: %#lx\n", shdr->sh_addr);
-		printf("sh_offset: %lu\n", shdr->sh_offset);
-		printf("sh_size: %lu\n", shdr->sh_size);
-		printf("sh_link: %u\n", shdr->sh_link);
-		printf("sh_info: %u\n", shdr->sh_info);
-		printf("sh_addralign: %lu\n", shdr->sh_addralign);
-		printf("sh_entsize: %lu\n", shdr->sh_entsize);
-
-		shdr = (void *)shdr + sizeof(Elf64_Shdr);
-	}
-}
-
-Elf64_Shdr				*get_section_bytype_64(Elf64_Ehdr *hdr, Elf64_Word type)
-{
-	Elf64_Shdr			*shdr;
-
-	shdr = (void *)hdr + hdr->e_shoff;
-	for (int i = 0; i < hdr->e_shnum; i++)
-	{
-		if (shdr->sh_type == type)
-			return (shdr);
-		shdr = (void *)shdr + sizeof(Elf64_Shdr);
-	}
-	return (shdr);
-}
-
-        /* .text:*/
-		/* sh_name: 157*/
-		/* sh_type: 1*/
-		/* sh_flags: 6*/
-		/* sh_addr: 0x580*/
-		/* sh_offset: 1408*/
-		/* sh_size: 450*/
-		/* sh_link: 0*/
-		/* sh_info: 0*/
-		/* sh_addralign: 16*/
-		/* sh_entsize: 0*/
-
 Elf64_Shdr				*add_shdr(void *packed, size_t section_size, size_t packed_size)
 {
 	Elf64_Ehdr			*hdr;
@@ -142,13 +51,23 @@ void					edit_phdr(void *packed, size_t section_size)
 {
 	Elf64_Phdr			*phdr;
 	Elf64_Ehdr			*hdr;
+	Elf64_Shdr			*bss;
 
 	hdr = packed;
 	phdr = (void *)hdr + hdr->e_phoff;
-	(void)section_size;
+	bss = get_section_bytype_64(hdr, SHT_NOBITS); // retrieve bss shdr to check which segment contains bss
+	// iterate through each program headers
 	for (int i = 0; i < hdr->e_phnum; i++)
 	{
-		phdr->p_flags = 7;
+		if (phdr->p_type == PT_LOAD) // if the current phdr is a loadable segment
+		{
+			if (phdr->p_vaddr <= bss->sh_addr && bss->sh_addr <= (phdr->p_vaddr + phdr->p_memsz)) // if the current segment contains the bss section
+			{
+				phdr->p_memsz += section_size; // increase it's virtual size by section_size
+				phdr->p_filesz += section_size;
+			}
+			phdr->p_flags = PF_X | PF_W | PF_R; // 7 or all permissions
+		}
 		phdr = (void *)phdr + sizeof(Elf64_Phdr);
 	}
 }
@@ -193,6 +112,7 @@ void					pack(void *m, struct stat *buf)
 
 	//
 	// inject code
+	ft_memcpy(new_sect, "AAAA", 4);
 	//
 
 	// add section header and update offsets
@@ -259,7 +179,7 @@ int					main(int ac, char **av)
 		return (1);
 	}
 
-	print_all(m);
+	print_phdr(m);
 	// Begin code injection
 	pack(m, &buf);
